test(ricer_test_tools)!: Move `stubs` module into `ricer_test_tools` lib
